requirements.txt
python-telegram-bot[webhooks]==20.3
python-dotenv==1.0.0
httpx==0.24.1
pyppeteer==1.0.2
